@@ -2,23 +2,11 @@
   "name": "elocyclo",
   "version": "0.1.0",
   "private": false,
-  "engines": {
-    "node": "12.18.3"
-  },
   "dependencies": {
-    "next": "^9.3.2",
-    "prop-types": "^15.5.10",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
-  },
-  "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^4.18.2",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-config-vogelint-react": "^1.0.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.1.0"
+    "next": "12.3.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "4.8.4"
   },
   "scripts": {
     "dev": "next",
