@@ -1,47 +1,38 @@
-import React, { Component } from 'react';
-// import PropTypes from 'prop-types';
+import { useCallback, useRef, useState } from "react";
 
-class SearchField extends Component {
-	constructor(props) {
-		super(props);
+function SearchField() {
+  const inputField = useRef();
+  const [opened, setIsOpened] = useState(false);
+  const [searchQuery, setSearchQuery] = useState(false);
 
-		this.state = { opened: false, searchQuery: '' };
-	}
-	setSearchFieldVisibility(opened) {
-		if (opened) {
-			this.inputField.focus();
-			this.inputField.select();
-		}
-		this.setState({ opened });
-	}
-	updateSearchQuery(searchQuery) {
-		this.setState({ searchQuery });
-	}
-	render() {
-		const { opened, searchQuery } = this.state;
-		return (
-			<div className={`search-button ${opened ? 'opened' : 'closed'}`}>
-				<small
-					className="simple-button gradient"
-					onClick={() => this.setSearchFieldVisibility(!opened)}
-				>
-					<b>{opened ? 'X' : 'ô'}</b>
-				</small>
-				<form>
-					<input
-						type="text"
-						value={searchQuery}
-						placeholder="Rechercher"
-						onChange={(evt) => this.updateSearchQuery(evt.target.value)}
-						ref={(input) => {
-							this.inputField = input;
-						}}
-					/>
-					<input type="submit" value="&ocirc;" />
-				</form>
-			</div>
-		);
-	}
+  const setSearchFieldVisibility = useCallback((newState) => {
+    if (newState && inputField.current) {
+      inputField.current.focus();
+      inputField.current.select();
+    }
+    setIsOpened(newState);
+  }, []);
+
+  return (
+    <div className={`search-button ${opened ? "opened" : "closed"}`}>
+      <small
+        className="simple-button gradient"
+        onClick={() => setSearchFieldVisibility(!opened)}
+      >
+        <b>{opened ? "X" : "ô"}</b>
+      </small>
+      <form>
+        <input
+          type="text"
+          value={searchQuery}
+          placeholder="Rechercher"
+          onChange={(evt) => setSearchQuery(evt.target.value)}
+          ref={inputField}
+        />
+        <input type="submit" value="&ocirc;" />
+      </form>
+    </div>
+  );
 }
 
 export default SearchField;
