@@ -1,18 +1,16 @@
-import React from 'react';
-
 export default () => (
-	<div id="slider-container" className="page-slider-container">
-		<section id="slider" className="page-slider">
-			<ul className="row">
-				<li>
-					<blockquote className="full">
-						<p>
-							<strong>Autant accessible que facile à utiliser!</strong>
-						</p>
-					</blockquote>
-				</li>
-			</ul>
-			<nav id="slider-nav" />
-		</section>
-	</div>
+  <div id="slider-container" className="page-slider-container">
+    <section id="slider" className="page-slider">
+      <ul className="row">
+        <li>
+          <blockquote className="full">
+            <p>
+              <strong>Autant accessible que facile à utiliser!</strong>
+            </p>
+          </blockquote>
+        </li>
+      </ul>
+      <nav id="slider-nav" />
+    </section>
+  </div>
 );
